Transferring existing application
This is the original MAGIC Helpdesk application developed by David Herbert in 2019.

As part of reorganising the Helpdesk into an issue tracking project and an implementation project, this code is being re-added to this repository and marked as a 1.0.0 release.

To allow it to be open-sourced sensitive properties have been redacted.