Adding new implemention
This commit represents a new version of this project, based on Microsoft 365 rather than a bespoke application as a backend.

It also seeks to unify all routes into the Helpdesk, such as email and phone/in-person, even if some are only documented rather than implemented here.

This new implementation is not yet complete, as the online form needs reviewing and a proper document library need setting up for storing file attachments for example.